--- src/main/java/com/sion/concertbooking/application/criteria/WaitingQueueCriteria.java
@@ -1,10 +1,0 @@
-package com.sion.concertbooking.application.criteria;
-
-import java.time.LocalDateTime;
-
-public record WaitingQueueCriteria(
-        long userId,
-        long concertId,
-        LocalDateTime registeredAt
-) {
-}
