+++ src/main/java/com/sion/concertbooking/application/payment/PaymentType.java
@@ -1,4 +1,4 @@
-package com.sion.concertbooking.application;
+package com.sion.concertbooking.application.payment;
 
 public enum PaymentType {
     FREE,
