+++ src/main/java/com/sion/concertbooking/application/point/PointChargeCriteria.java
@@ -1,6 +1,6 @@
-package com.sion.concertbooking.application.criteria;
+package com.sion.concertbooking.application.point;
 
-import com.sion.concertbooking.application.PaymentType;
+import com.sion.concertbooking.application.payment.PaymentType;
 
 public record PointChargeCriteria(
         long userId,
