+++ src/main/java/com/sion/concertbooking/application/point/PointDeductor.java
@@ -1,6 +1,5 @@
-package com.sion.concertbooking.application;
+package com.sion.concertbooking.application.point;
 
-import com.sion.concertbooking.application.result.PointResult;
 import com.sion.concertbooking.domain.pointhistory.PointHistoryInfo;
 import com.sion.concertbooking.domain.point.PointInfo;
 import com.sion.concertbooking.domain.pointhistory.PointHistoryService;
