+++ src/main/java/com/sion/concertbooking/intefaces/presentation/response/PointChargeResponse.java
@@ -2,8 +2,8 @@ package com.sion.concertbooking.intefaces.presentation.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sion.concertbooking.application.PaymentType;
-import com.sion.concertbooking.application.result.PointResult;
+import com.sion.concertbooking.application.payment.PaymentType;
+import com.sion.concertbooking.application.point.PointResult;
 import com.sion.concertbooking.domain.pointhistory.TransactionType;
 
 import java.time.LocalDateTime;
