+++ src/main/java/com/sion/concertbooking/intefaces/presentation/response/PointUseResponse.java
@@ -2,7 +2,7 @@ package com.sion.concertbooking.intefaces.presentation.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sion.concertbooking.application.result.PointResult;
+import com.sion.concertbooking.application.point.PointResult;
 import com.sion.concertbooking.domain.pointhistory.TransactionType;
 
 import java.time.LocalDateTime;
