+++ src/test/java/com/sion/concertbooking/application/PointChargerTest.java
@@ -1,5 +1,7 @@
 package com.sion.concertbooking.application;
 
+import com.sion.concertbooking.application.payment.PaymentType;
+import com.sion.concertbooking.application.point.PointCharger;
 import com.sion.concertbooking.domain.point.PointInfo;
 import com.sion.concertbooking.domain.point.PointService;
 import com.sion.concertbooking.domain.pointhistory.PointHistoryService;
